adding excel_helper and moving main logic to spotify.py file
