+++ requirements.txt
@@ -1,2 +1,3 @@
 python-dotenv==1.0.0
 spotipy==2.23.0
+openpyxl==3.1.2
